[CEDS-2961] Content update of the Additional Document Codes page
Adding two new expanders, new content and some dynamic
content (with dynamic links to tariff guide).
