+++ conf/application.conf
@@ -160,6 +160,7 @@ urls {
   serviceAvailability = "https://www.gov.uk/guidance/customs-declaration-service-service-availability-and-issues"
   customsMovementsFrontend = "http://localhost:6796/customs-movements"
   tradeTariff = "https://www.gov.uk/trade-tariff"
+  tradeTariffSections = "https://www.trade-tariff.service.gov.uk/sections"
   tariffCommodities = "https://www.trade-tariff.service.gov.uk/commodities/"
   previousProcedureCodes = "https://www.gov.uk/government/publications/appendix-1-de-110-requested-and-previous-procedure-codes"
   tradeTariffVol3ForCds2 = "https://www.gov.uk/government/collections/uk-trade-tariff-volume-3-for-cds--2"
@@ -186,6 +187,10 @@ urls {
   goodsLocationCodesForDataElement = "https://www.gov.uk/government/collections/goods-location-codes-for-data-element-523-of-the-customs-declaration-service"
   tariffCdsChiefSupplement = "https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/956071/Tariff_Volume_3_Great_Britain_Supplement_CDS_and_CHIEF_version_5.3__1_.odt"
   notesForMucrConsolidation = "https://www.gov.uk/government/publications/uk-trade-tariff-cds-volume-3-export-declaration-completion-guide/group-2-references-of-messages-document-certificates-and-authorisations#de-21-simplified-declaration-previous-documents-box-40-declaration-previous-document"
+  additionalDocumentsLicenceTypes = "https://www.gov.uk/guidance/data-element-23-document-and-other-reference-codes-licence-types-imports-and-exports-of-the-customs-declaration-service-cds"
+  additionalDocumentsUnionCodes = "https://www.gov.uk/government/publications/data-element-23-documents-and-other-reference-codes-union-of-the-customs-declaration-service-cds"
+  additionalDocumentsReferenceCodes = "https://www.gov.uk/guidance/data-element-23-documents-and-other-reference-codes-national-of-the-customs-declaration-service-cds#table-2-national-document-codes"
+  additionalDocumentsUnitCodes = "https://www.gov.uk/government/publications/measurement-unit-codes-for-data-elements-23-and-44-of-the-customs-declaration-service"
 
   procedureCodes {
     removalOfGoodsFromExciseWarehouse = "https://www.gov.uk/guidance/receive-goods-into-and-remove-goods-from-an-excise-warehouse-excise-notice-197"
